moved the tests for votecast into their own file.
